♻️[refactor] 자산 조회-등록-상세 로직 수정
